Add support for Git dirty-checking
Sometimes, all you care about is whether or not a Git repository is
dirty, or changed, at all. This change introduces the concept of a dirty
Git repository as one that either has conflicts or has untracked, modified,
staged, renamed, or deleted files.

Since Starship currently calculates the repository status in total
instead of lazily calculating the values used in the prompt, this change
sets the flag at each point that it can. This adds a negligible amount
of work and ensures we catch each of the statuses that "dirty"
coalesces.

The tests mostly form regression coverage since the status is simpler
than its sibling statuses. Since it is only a unidirectional flag
status, there is no test for the state where a repository meets multiple
conditions.